chore(constants): move Family, Gender to global constants
